[spatialite] Move gui related provider code to a new file
